+++ crates/dioxus-class/src/components/mod.rs
@@ -1,9 +1,0 @@
-pub mod item;
-pub mod list;
-
-pub mod prelude {
-    pub use super::item::ItemProps;
-    pub use super::list::ListProps;
-    pub use super::list::List;
-    pub use super::list::ListC;
-}