+++ crates/dioxus-class/src/lib.rs
@@ -7,15 +7,11 @@ pub mod classes;
 
 pub mod macros;
 
-pub mod components;
-
 pub mod build;
 
 pub mod prelude {
     pub use crate::class::Class;
     pub use crate::classes::Classes;
-
-    pub use crate::components::prelude::*;
 
     pub use crate::class;
     pub use crate::style_type;