+++ crates/dioxus-daisyui/src/components/menu.rs
@@ -1,6 +1,0 @@
-use dioxus::prelude::*;
-use crate::prelude::*;
-
-pub fn Menu<'a>(cx: Scope<'a, ListProps<'a>>) -> Element {
-    ListC(cx, class!(menu))
-}