+++ demo/emoji-browser/src/main.rs
@@ -1,0 +1,5 @@
+use emoji_browser::app::App;
+
+fn main() {
+    dioxus_web::launch(App);
+}