+++ demo/emoji-browser/static/css/tailwind.config.js
@@ -1,0 +1,12 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+  content: [
+    "../pages/*.html",
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [
+    require("daisyui")
+  ],
+}