+++ crates/dioxus-tailwindcss/CHANGELOG.md
@@ -9,7 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- add build::tailwindcss to run tailwindcss process
+
 ### Fixed
+
+## [0.2.2] - 2023-02-27
+
+### Fixed
+
+- cleanup
 
 ## [0.2.1] - 2023-02-27